+++ cli/CHANGELOG.md
@@ -5,7 +5,7 @@ _Released 12/19/2023 (PENDING)_
 
 **Bugfixes:**
 
-- Fixed asset capture for Test Replay (which will potentially affect styles) specifically for requests that are routed through service workers. Fixes [#28516](https://github.com/cypress-io/cypress/issues/28516).
+- Fixed asset capture for Test Replay for requests that are routed through service workers. This addresses an issue where styles were not being applied properly in Test Replay. Fixes [#28516](https://github.com/cypress-io/cypress/issues/28516).
 - Cypress no longer errors during `cypress run` or `cypress open` when using Typescript 5.3.2+ with `extends` in `tsconfig.json`. Fixes [#28385](https://github.com/cypress-io/cypress/issues/28385).
 - Fixed a regression in [`13.6.1`](https://docs.cypress.io/guides/references/changelog/13.6.1) where a malformed URI would crash Cypress. Fixes [#28521](https://github.com/cypress-io/cypress/issues/28521).
 - Fixed a regression in [`12.4.0`](https://docs.cypress.io/guides/references/changelog/12.4.0) where erroneous `<br>` tags were displaying in error messages in the Command Log making them less readable. Fixes [#28452](https://github.com/cypress-io/cypress/issues/28452).
