+++ packages/proxy/lib/http/util/headers.ts
@@ -1,0 +1,3 @@
+export const hasServiceWorkerHeader = (headers: Record<string, string | string[] | undefined>) => {
+  return headers?.['service-worker'] === 'script' || headers?.['Service-Worker'] === 'script'
+}
